Bug fixing (#40)
* Fix various things and refactoring

* Fixing CI to limit the testing to `test_*.py` files.

* Removing `cfg.checkpoints_dir` because we have been only using `cfg.model` in benchmark scripts for either huggingface repo or direct path.

* fix bad import and a bug in image task.

* fix bad import and a bug in image task.

* Evaluate in separate threads.

* Prevent warnings from tokenizer by disabling parallelism (we don't have a large dataset to tokenize).

* Revert the shebang. Better not to have a shebang for potential env problems.
