+++ src/openelm/config/elm_sodarace_cfg.yaml
@@ -1,5 +1,4 @@
 model: codegen-350M-mono
-checkpoints_dir: checkpoints
 cuda: True
 gpus: 1
 seed: 42
